refactor(forum_moderation): déplacement d'enum dans un fichier enums.py
