+++ lacommunaute/forum_moderation/views.py
@@ -5,8 +5,8 @@ from machina.apps.forum_moderation.views import (
     TopicDeleteView as BaseTopicDeleteView,
 )
 
+from lacommunaute.forum_moderation.enums import BlockedPostReason
 from lacommunaute.forum_moderation.models import BlockedEmail, BlockedPost
-from lacommunaute.forum_moderation.utils import BlockedPostReason
 
 
 class TopicDeleteView(BaseTopicDeleteView):
